+++ Service/INavigationService.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TODO.Service
+{
+    public interface INavigationService
+    {
+        Core.AbstractViewMode CurrentView { get; }
+        void NavigateTo<T>() where T : Core.AbstractViewMode;
+
+    }
+}
