+++ View/HomeView.xaml.cs
@@ -5,7 +5,7 @@ namespace TODO.View
     /// <summary>
     /// Interaction logic for HomeView.xaml
     /// </summary>
-    public partial class HomeView : UserControl
+    public partial class HomeView
     {
         public HomeView()
         {
